+++ src/components/Ideas.vue
@@ -6,20 +6,19 @@ import Idea from './Idea.vue'
 <template>
   <aside class="mb-5 p-5 border-4 border-sky-500">
     <SectionTitle anchor="ideas" title="Ideas" />
-    <Idea
-      >Keywords section with 'Living and working in Antarctica' and 'Topographic mapping'
-      options.</Idea
-    >
-    <Idea>Citation section related to DOI.</Idea>
+    <Idea>BAS Research Topics topic keywords.</Idea>
+    <Idea>ISO topic mapped from research topic.</Idea>
     <Idea>Place keywords based on Well-Known Extents.</Idea>
+    <Idea complete>Table of contents.</Idea>
+    <Idea complete>Dark mode.</Idea>
+    <Idea complete>DOI identifier based on file identifier.</Idea>
+    <Idea complete>Esri identifier.</Idea>
+    <Idea>Citation.</Idea>
+    <Idea>Citation based on DOI.</Idea>
     <Idea>Include `protocol: 'WWW:LINK-1.0-http--link'` in download distribution options.</Idea>
     <Idea>Include WMTS, WFS and Esri in services section.</Idea>
     <Idea>Download proxy items.</Idea>
     <Idea>Colour code distribution options like JWTs.</Idea>
-    <Idea complete>Table of contents.</Idea>
-    <Idea complete>Dark mode.</Idea>
-    <Idea>DOI identifier based on file identifier.</Idea>
-    <Idea>Esri identifier.</Idea>
     <p class="mt-4 text-black dark:text-white">
       <em>If there's something else you think would be a good idea let @felnne know.</em>
     </p>